+++ lib/article_json/export/plain_text/elements/embed.rb
@@ -1,0 +1,16 @@
+module ArticleJSON
+  module Export
+    module PlainText
+      module Elements
+        class Embed < Base
+          # Embedded elements are being skipped when rendering plain text.
+          # Therefore, return an empty string.
+          # @return [String]
+          def export
+            ''
+          end
+        end
+      end
+    end
+  end
+end