+++ lib/article_json/export/plain_text/elements/text_box.rb
@@ -1,0 +1,16 @@
+module ArticleJSON
+  module Export
+    module PlainText
+      module Elements
+        class TextBox < Base
+          # Text boxes are being skipped when rendering plain text. Therefore,
+          # return an empty string.
+          # @return [String]
+          def export
+            ''
+          end
+        end
+      end
+    end
+  end
+end